% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_filter.R
\name{add_filter}
\alias{add_filter}
\title{Acronyms and Glossaries support in RMarkdown}
\usage{
add_filter(error = TRUE)
}
\arguments{
\item{error}{Whether to raise an error if the required version of pandoc
is not found}
}
\description{
This method allows to easily register the 'parse-acronyms' Lua filter
in a RMarkdown document.
This filter is ultimately responsible for every functionality of this
package: loading the list of acronyms, creating the glossary in the
document, and linking acronyms to their definition.
}
\details{
It should be used inside the YAML metadata of a RMarkdown document.
For example,
\preformatted{
---
output:
    html_document:
    pandoc_args: !expr acronymsdown::add_filter()
---
}

This method is largely inspired from 'rmdfiltr'
(https://github.com/crsh/rmdfiltr/)
}
