% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_format.R
\name{md_format}
\alias{md_format}
\title{Sample output format for Acronymsdown documents}
\usage{
md_format(...)
}
\arguments{
\item{...}{Additional arguments, which will be passed to the base
\code{\link[rmarkdown:md_document]{rmarkdown::md_document()}} format.}
}
\value{
A rmarkdown output format, which can be used in
\code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Format for converting R Markdown documents to (pandoc) markdown documents.
This format is primarily used for testing purposes.
}
\details{
This output format was meant for testing the Acronymsdown package.
You may use it in your documents, but be warned that there are some
opinionated choices. In particular, this format:
\itemize{
\item ensures that headings use the ATX syntax (e.g., \verb{# Title}, \verb{## Subtitle}) ;
\item preserves line wrappings ;
\item forces the pandoc variant of \emph{markdown} (instead of the default,
strict markdown) ;
\item includes the \emph{acronymsdown} Lua Filter.
}

These choices make sense in order to facilitate the automated testing
of Acronymsdown, by ensuring a reliable (and reproducible) result,
even for older versions of pandoc.
}
